roadmap: Add debug mode, cache truncated text.
Done for performance improvement.
