+++ cmd/roadmap/webapp/.env.development
@@ -2,3 +2,4 @@
 # `npm run dev`). For local-only overrides create a `.env.local`.
 
 VITE_API_URL=http://localhost:8080
+VITE_DEBUG_RENDER=true
