+++ cmd/roadmap/webapp/timeline.ts
@@ -2,6 +2,7 @@ import p5 from "p5";
 
 import { fetchTasks, submitTasks } from "./data";
 import { getOwnerColor, getPriorityColor } from "./config";
+import { IDEA_STATUS, HALTED_STATUS, isIdeaTask, isHaltedTask, isSidebarTask, getDateBounds, hexToP5Color, drawArrow, truncateText } from "./helpers";
 
 let tasks: any[] = [];
 let timelineTasks: any[] = [];
@@ -26,8 +27,7 @@ let ideasSidebarCollapsed = false;
 let ideaSectionCollapsed = false;
 let haltedSectionCollapsed = false;
 
-const IDEA_STATUS = "idea";
-const HALTED_STATUS = "progress made, halted";
+
 const sidebarExpandedWidth = 380;
 const sidebarCollapsedWidth = 36;
 const sidebarPadding = 12;
@@ -38,13 +38,13 @@ const sidebarCardGap = 8;
 type SidebarSectionKey = "ideas" | "halted";
 type SidebarHit =
   | {
-      kind: "section";
-      section: SidebarSectionKey;
-    }
+    kind: "section";
+    section: SidebarSectionKey;
+  }
   | {
-      kind: "card";
-      task: any;
-    };
+    kind: "card";
+    task: any;
+  };
 
 // p5.js sketch
 const sketch = (p: p5) => {
@@ -285,6 +285,10 @@ const sketch = (p: p5) => {
     }
     if (!isDragging && !redraw) {
       return;
+    }
+
+    if (import.meta.env.VITE_DEBUG_RENDER === "true") {
+      console.log("🎨 Rendering frame at", new Date().toISOString());
     }
 
     // console.log("drawing timeline...");
@@ -693,39 +697,7 @@ const sketch = (p: p5) => {
 
 new p5(sketch);
 
-function isIdeaTask(task: any): boolean {
-  return (
-    String(task.status || "")
-      .trim()
-      .toLowerCase() === IDEA_STATUS
-  );
-}
 
-function isHaltedTask(task: any): boolean {
-  return (
-    String(task.status || "")
-      .trim()
-      .toLowerCase() === HALTED_STATUS
-  );
-}
-
-function isSidebarTask(task: any): boolean {
-  return isIdeaTask(task) || isHaltedTask(task);
-}
-
-function getDateBounds(items: any[]): { start: number; end: number } | null {
-  const starts = items.map((task) => new Date(task.start).getTime()).filter((value) => !isNaN(value));
-  const ends = items.map((task) => new Date(task.end).getTime()).filter((value) => !isNaN(value));
-
-  if (starts.length === 0 || ends.length === 0) {
-    return null;
-  }
-
-  return {
-    start: Math.min(...starts),
-    end: Math.max(...ends),
-  };
-}
 
 function setTimelineRange(items: any[]) {
   const bounds = getDateBounds(items);
@@ -862,40 +834,7 @@ function xToDate(x: number, p: p5): string {
   return new Date(dateMillis).toISOString().split("T")[0]; // Format as YYYY-MM-DD
 }
 
-function hexToP5Color(hex: string, alpha: number, p: p5) {
-  let col = p.color(hex); // Convert hex to p5 color
-  col.setAlpha(alpha * 255); // p5.js alpha is 0-255, so multiply by 255
-  return col;
-}
 
-function drawArrow(p: p5, x1: number, y1: number, x2: number, y2: number) {
-  p.stroke(100);
-  p.strokeWeight(2);
-  p.noFill();
-
-  let midX = (x1 + x2) / 2;
-  let midY = (y1 + y2) / 2;
-
-  // Draw a curved line (Bezier)
-  p.beginShape();
-  p.vertex(x2, y2); // Swapped start & end
-  p.bezierVertex(midX, y2, midX, y1, x1, y1); // Flipped Bezier direction
-  p.endShape();
-
-  // Move the arrowhead to `x1, y1` (flipped direction)
-  let arrowSize = 6;
-  let angle = Math.atan2(y1 - y2, x1 - x2); // Flip angle calculation
-  p.fill(100);
-  p.noStroke();
-  p.triangle(
-    x1,
-    y1, // Arrowhead now correctly at the **start** of the line
-    x1 - arrowSize * Math.cos(angle + Math.PI / 6),
-    y1 - arrowSize * Math.sin(angle + Math.PI / 6),
-    x1 - arrowSize * Math.cos(angle - Math.PI / 6),
-    y1 - arrowSize * Math.sin(angle - Math.PI / 6),
-  );
-}
 
 function drawStatusSidebar(p: p5) {
   const sidebarWidth = getIdeasSidebarWidth(p);
@@ -1053,15 +992,7 @@ function drawIdeasSidebarToggle(p: p5, sidebarX: number) {
   p.text(ideasSidebarCollapsed ? "<" : ">", toggleX + toggleSize / 2, toggleY + toggleSize / 2 - 1);
 }
 
-function truncateText(p: p5, value: string, maxWidth: number): string {
-  if (p.textWidth(value) <= maxWidth) return value;
 
-  let truncated = value;
-  while (truncated.length > 0 && p.textWidth(`${truncated}...`) > maxWidth) {
-    truncated = truncated.slice(0, -1);
-  }
-  return `${truncated}...`;
-}
 
 function drawTooltip(p: p5, task: any, x: number, y: number) {
   const padding = 8;
